Restructured and made asynch

+++ src/main/java/org/jboss/infinispan/jazz/test/utils/KeyValueGenerator.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 public class KeyValueGenerator {
 
-    public static Map<String, String> generateKeyValues(Long count, Long keyLength, Long valueLength) {
+    public static Map<String, String> generateKeyValues(int count, int keyLength, int valueLength) {
         Map<String, String> result = new HashMap<>();
         while (result.size() < count) {
             result.put(StringGenerator.value(keyLength), StringGenerator.value(valueLength));
